Moving all utils into the DicomImageRedactorEngine class
