--- presidio-image-redactor/presidio_image_redactor/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Image Redactor utilities."""
